chore(posters): rename and reorganize carousel poster handler
