+++ src/handlers/poster/hubble/components/OverlayHubble.tsx
@@ -1,4 +1,4 @@
-import { OverlayProps } from '../../PosterBaseView';
+import { OverlayProps } from '../../CarouselPosterView.tsx';
 import HubbleProgressBar from './HubbleProgressBar';
 export default function OverlayHubble({ poster, seconds, posterIndex, nextPoster, pausePoster }: OverlayProps) {
   return (
