+++ src/handlers/poster/hubble/index.tsx
@@ -1,12 +1,12 @@
 import './components/index.scss';
 import { RequestResult } from '@hey-api/client-fetch';
-import PosterBaseView from '../PosterBaseView';
+import CarouselPosterView from '../CarouselPosterView.tsx';
 import { BasePosterResponse, getHubblePosters } from '../../../api';
 import OverlayHubble from './components/OverlayHubble';
 
 export default function PosterHubbleView() {
   return (
-    <PosterBaseView
+    <CarouselPosterView
       overlay={OverlayHubble}
       getPosters={getHubblePosters as unknown as () => Promise<RequestResult<BasePosterResponse>>}
     />
