+++ src/handlers/poster/types/LogoPoster.tsx
