+++ src/handlers/poster/types/TrainPoster.scss
