+++ src/handlers/poster/types/VideoPoster.tsx
